Jwt token implementation

+++ src/main/java/com/example/library/management/controller/UserController.java
@@ -1,16 +1,13 @@
 package com.example.library.management.controller;
 
-import com.example.library.management.dto.UserRequest;
 import com.example.library.management.dto.UserResponse;
+import com.example.library.management.security.AccessEnforcer;
 import com.example.library.management.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -21,34 +18,24 @@ public class UserController
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
     private final UserService userService;
+    private final AccessEnforcer accessEnforcer;
 
-    public UserController(UserService userService) {
+    public UserController(UserService userService, AccessEnforcer accessEnforcer) {
         this.userService = userService;
+        this.accessEnforcer = accessEnforcer;
     }
 
     @GetMapping("/users")
     public List<UserResponse> getAllUsers() {
-        log.debug("Received get all users request");
+        accessEnforcer.requireAdmin();
 
         return userService.getAllUsers();
     }
 
-    @PostMapping("/users")
-    public UserResponse createUser(@RequestBody UserRequest userRequest) {
-        log.debug("Received create user request for email={}", userRequest.email());
-
-        return userService.createUser(userRequest);
-    }
-
-    @PostMapping("/users/login")
-    public ResponseEntity<UserResponse> loginUser(@RequestBody UserRequest loginDto) {
-        log.debug("Received user login request for email={}", loginDto.email());
-
-        return userService.getUser(loginDto.email(), loginDto.password());
-    }
-
     @DeleteMapping("/users/{userId}")
     public void deleteUser(@PathVariable Long userId) {
+        accessEnforcer.requireAdmin();
+
         log.debug("Received delete user request for userId={}", userId);
 
         userService.deleteUser(userId);
